% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_estimator.R
\name{iv_sim}
\alias{iv_sim}
\title{Simulate a dataset and compute IV.}
\usage{
iv_sim(G, N_target)
}
\arguments{
\item{N_target}{}
}
\value{
$hat{theta}_{IV}-theta_{IV}$
}
\description{
This function (1) Simulates the data. (2) Computes the IV coefficient and
(3) Outputs the difference between the estimated IV and the truth.
}
\details{
#'
}
\examples{
iv_sim(235,30000)
}
