% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{npiv_cbar_fn}
\alias{npiv_cbar_fn}
\title{This calculates a smoothed coefficient semiparametric IV estimator.}
\usage{
npiv_cbar_fn(cbar, h, Y, X, C, M)
}
\arguments{
\item{M}{}
}
\description{
This calculates a smoothed coefficient semiparametric IV estimator.
}
