% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{npiv_fn}
\alias{npiv_fn}
\title{Compute average treatment effects of the model, where the localize step
computes a semiparametric smoothed coefficient IV estimator.}
\usage{
npiv_fn(h, Y, X, C, D, Z, C_g, rho_g, M)
}
\arguments{
\item{M}{}
}
\description{
Compute average treatment effects of the model, where the localize step
computes a semiparametric smoothed coefficient IV estimator.
}
