% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_estimator.R
\name{rcestimator_sim}
\alias{rcestimator_sim}
\title{Simulate a dataset and compute the random coefficients estimator.}
\usage{
rcestimator_sim(G, N_target, h)
}
\arguments{
\item{h}{}
}
\description{
This function (1) Simulates the data. (2) Computes the random coefficient
(RC) estimator and
(3) Outputs the difference between the RC estimate and the truth.
}
