% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{thetahat_cbar_fn}
\alias{thetahat_cbar_fn}
\title{Kernel estimation of a varying coefficients model of a model}
\usage{
thetahat_cbar_fn(cbar, h, Y, W, C)
}
\arguments{
\item{cbar}{A specific (non-random) point at which to evaluate $theta(r)$.}

\item{h}{The selected bandwidth.}

\item{Y}{The outcome variable.}

\item{W}{The regressor variable.}

\item{C}{The variable that determines the coeffients. In our case this is the
share of compliers.}
}
\description{
$$ Y_{ig} = W_{ig}'theta(C_g) + U_{ig}$$
}
\details{
where $U_{ig} = W_{ig}'(Theta_{ig} - E[Theta_{ig}|C_g])

At present the function implements the uniform kernel. This functions estimates
$$ theta(cbar) = E[theta(C_g) | C_g = cbar]$$.

The function multiplies each row of the data times a kernel weight. With these
transformed variables we compute an OLS regression.
}
