% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{thetahat_fn}
\alias{thetahat_fn}
\title{Compute average treatment effects of the model}
\usage{
thetahat_fn(h, Y, W, C, D, Z, C_g, rho_g)
}
\arguments{
\item{h}{Bandwidth choice to compute $theta(R_g)$.}

\item{Y}{Outcome variable.}

\item{W}{Set of regressors.}

\item{C}{The variable that determines the coeffients. In our case this is the
share of compliers.}

\item{D}{An individual take-up variable.}

\item{Z}{An individual offer variable.}

\item{C_g}{A vector with the share of compliers in each city.}

\item{rho_g}{A vector with the ratio of city size to the average city size.}
}
\description{
$$ Y_{ig} = W_{ig}'theta(C_g) + U_{ig}$$
}
\details{
where $U_{ig} = W_{ig}'(Theta_{ig} - E[Theta_{ig}|C_g])$.

This function computes an estimate for $theta = E[theta(R_g)]$.
The function procedes in three steps: (1) For each city compute the estimated
coefficients $hat{theta}(R_g)$. (2) Compute a weighted averagem of the estimated
coeffients, depending on population size. (3) Adjust coefficients for the direct
and interaction effects to account for the total share of compliers.

Notice that the kernel function is constructed with the whole data that
has individuals in each row rather than cities. We could do step (1) and (2)
in a single step by averaging over individuals rather than computing a weighted
average over cities. However, doing it this way is faster because we only need
to compute a kernel regression for each city and weight by the population in
that city. Otherwise we would need to repeat a lot of computations for individuals
in the same city.
}
